refactor(sse): extract combo 400-classification predicates into a leaf module
combo.ts sits at 3640 lines against a frozen file-size ratchet of 3642. The gate is
shrink-only, so ANY change that adds a line to this file is rejected -- including a
one-line bug fix plus its explanatory comment. Two in-flight #7847 fixes both hit this,
and one shipped a violation because lint-staged's Prettier reflowed an unrelated over-long
line (+4 lines) AFTER the gate had already been checked.

Squeezing comments to fit is optimizing the wrong thing. This restores real headroom
instead: the three 400-classification predicates are pure string classifiers over upstream
error text, sharing the canonical pattern lists in accountFallback.ts, and have no business
inside the routing loop. Moving them takes combo.ts to 3610 -- 30 lines of headroom.

Behavior-preserving. The bodies were extracted from both revisions and compared after
normalizing whitespace: 3 of 3 byte-identical. combo.ts imports them for the fallback guard
and re-exports them, so the public surface is unchanged and
tests/unit/combo-param-validation-fallback-4519.test.ts still imports them from combo.ts
without modification (6/6 pass).

Note for anyone repeating this: a bare re-export does not bind the name locally, so
combo.ts's own use of the predicates needed a separate import. typecheck caught it.
